+++ packages/outliner/src/index.tsx
@@ -19,6 +19,7 @@ export function Outliner({
 	hiddenRowIds,
 	completionExitRowIds,
 	contextRowIds,
+	noVisibleChildrenRowIds,
 	newNodeDueDate,
 	newNodeTags,
 	existingTags,
@@ -46,6 +47,7 @@ export function Outliner({
 			hiddenRowIds={hiddenRowIds}
 			completionExitRowIds={completionExitRowIds}
 			contextRowIds={contextRowIds}
+			noVisibleChildrenRowIds={noVisibleChildrenRowIds}
 			existingTags={existingTags}
 			onDeleteTag={onDeleteTag}
 			onTagClick={onTagClick}